+++ cwlogs/__init__.py
@@ -11,7 +11,7 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-__version__ = '1.3.2'
+__version__ = '1.3.3'
 
 
 def awscli_initialize(cli):